Start CD cleaning
