+++ includes/exec.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exec.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbrucker <sbrucker@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/23 13:04:09 by sbrucker          #+#    #+#             */
-/*   Updated: 2018/10/30 15:44:54 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/11/17 12:23:53 by cgaspart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,25 +64,6 @@ t_exec	*create_exec(t_environ *env);
 // Path parsing
 
 char	*isin_path(char *path_entry, char *cmd);
-
-// Builtins
-
-void	builtin_exit(char **argv, t_environ *env, t_exec *exe);
-void	builtin_echo(char **argv, t_environ *env, t_exec *exe);
-void	builtin_env(char **argv, t_environ *env, t_exec *exe);
-void	builtin_setenv(char **argv, t_environ *env, t_exec *exe);
-void	builtin_unsetenv(char **argv, t_environ *env, t_exec *exe);
-void	builtin_cd(char **argv, t_environ *env, t_exec *exe);
-void	builtin_return(char **argv, t_environ *env, t_exec *exe);
-void	builtin_toggle_syntax_highlighting(char **argv, t_environ *env, \
-			t_exec *exe);
-void	builtin_test(char **argv, t_environ *env, t_exec *exe);
-void	builtin_history(char **argv, t_environ *env, t_exec *exe);
-void	builtin_toggle_syntax_highlighting(char **argv, t_environ *env,
-			t_exec *exe);
-void	builtin_alias(char **argv, t_environ *env, t_exec *exe);
-void	builtin_unalias(char **argv, t_environ *env, t_exec *exe);
-
 char	*get_env(const char *name, const char **envp);
 int		get_env_pos(const char *name, const char **envp);
