+++ srcs/autocomplete/autoc_dir.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   autoc_dir.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cyfermie <cyfermie@student.42.fr>          +#+  +:+       +#+        */
+/*   By: cgaspart <cgaspart@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/29 17:18:05 by cgaspart          #+#    #+#             */
-/*   Updated: 2018/10/19 16:59:08 by cyfermie         ###   ########.fr       */
+/*   Updated: 2018/11/17 12:33:27 by cgaspart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
